+++ test/lib/absinthe/schema_test.exs
@@ -13,7 +13,7 @@ defmodule Absinthe.SchemaTest do
     end
 
     it "are loaded" do
-      load_valid_schema
+      load_valid_schema()
       assert map_size(Absinthe.Type.BuiltIns.__absinthe_types__) > 0
       Absinthe.Type.BuiltIns.__absinthe_types__
       |> Enum.each(fn
@@ -22,7 +22,7 @@ defmodule Absinthe.SchemaTest do
       end)
       int = ValidSchema.__absinthe_type__(:integer)
       assert 1 == Type.Scalar.serialize(int, 1)
-      assert {:ok, 1} == Type.Scalar.parse(int, 1)
+      assert {:ok, 1} == Type.Scalar.parse(int, 1, %{})
     end
 
   end